+++ packages/experimental/test/Ndjson.test.ts
@@ -23,7 +23,7 @@ describe("Ndjson", () => {
   test("socket", () =>
     Effect.gen(function*(_) {
       const socket = Socket.makeNetChannel<Ndjson.NdjsonError>({ port, host: "localhost" }).pipe(
-        Ndjson.duplex
+        Ndjson.duplex()
       )
 
       const outputEffect = Stream.make({ hello: "world" }, { test: 123 }).pipe(
@@ -38,7 +38,7 @@ describe("Ndjson", () => {
   test("socket x10000", () =>
     Effect.gen(function*(_) {
       const socket = Socket.makeNetChannel<Ndjson.NdjsonError>({ port }).pipe(
-        Ndjson.duplex
+        Ndjson.duplex()
       )
       const msgs = Array.from({ length: 10000 }, (_, i) => ({ hello: i }))
 
@@ -49,5 +49,49 @@ describe("Ndjson", () => {
       const output = yield* _(outputEffect)
 
       assert.deepStrictEqual(Chunk.toArray(output), msgs)
+    }).pipe(Effect.runPromise))
+
+  test("should ignore empty lines", () =>
+    Effect.gen(function*() {
+      const encoder = new TextEncoder()
+
+      const ndjson = [
+        "{\"id\":\"1\"}",
+        "{\"id\":\"2\"}",
+        "\n",
+        "{\"id\":\"3\"}",
+        "{\"id\":\"4\"}"
+      ].join("\n")
+
+      const results = yield* Stream.succeed(encoder.encode(ndjson)).pipe(
+        Stream.pipeThroughChannel(Ndjson.unpack({ ignoreEmptyLines: true })),
+        Stream.runCollect,
+        Effect.map(Chunk.toReadonlyArray)
+      )
+
+      assert.deepStrictEqual(results, [{ id: "1" }, { id: "2" }, { id: "3" }, { id: "4" }])
+    }).pipe(Effect.runPromise))
+
+  test("should not ignore empty lines", () =>
+    Effect.gen(function*() {
+      const encoder = new TextEncoder()
+
+      const ndjson = [
+        "{\"id\":\"1\"}",
+        "{\"id\":\"2\"}",
+        "\n",
+        "{\"id\":\"3\"}",
+        "{\"id\":\"4\"}"
+      ].join("\n")
+
+      const error = yield* Stream.succeed(encoder.encode(ndjson)).pipe(
+        Stream.pipeThroughChannel(Ndjson.unpack()),
+        Stream.runCollect,
+        Effect.map(Chunk.toReadonlyArray),
+        Effect.flip
+      )
+
+      assert.instanceOf(error, Ndjson.NdjsonError)
+      assert.propertyVal(error, "reason", "Unpack")
     }).pipe(Effect.runPromise))
 })