+++ packages/platform/src/internal/httpClientResponse.ts
@@ -67,19 +67,19 @@ class ClientResponseImpl extends Inspectable.Class implements ClientResponse.Htt
 
   get stream(): Stream.Stream<Uint8Array, Error.ResponseError> {
     return this.source.body
-      ? Stream.fromReadableStream(() => this.source.body!, (_) =>
+      ? Stream.fromReadableStream(() => this.source.body!, (cause) =>
         new Error.ResponseError({
           request: this.request,
           response: this,
           reason: "Decode",
-          error: _
+          cause
         }))
       : Stream.fail(
         new Error.ResponseError({
           request: this.request,
           response: this,
           reason: "EmptyBody",
-          error: "can not create stream from empty body"
+          description: "can not create stream from empty body"
         })
       )
   }
@@ -87,12 +87,12 @@ class ClientResponseImpl extends Inspectable.Class implements ClientResponse.Htt
   get json(): Effect.Effect<unknown, Error.ResponseError> {
     return Effect.tryMap(this.text, {
       try: (text) => text === "" ? null : JSON.parse(text) as unknown,
-      catch: (_) =>
+      catch: (cause) =>
         new Error.ResponseError({
           request: this.request,
           response: this,
           reason: "Decode",
-          error: _
+          cause
         })
     })
   }
@@ -101,12 +101,12 @@ class ClientResponseImpl extends Inspectable.Class implements ClientResponse.Htt
   get text(): Effect.Effect<string, Error.ResponseError> {
     return this.textBody ??= Effect.tryPromise({
       try: () => this.source.text(),
-      catch: (_) =>
+      catch: (cause) =>
         new Error.ResponseError({
           request: this.request,
           response: this,
           reason: "Decode",
-          error: _
+          cause
         })
     }).pipe(Effect.cached, Effect.runSync)
   }
@@ -115,12 +115,12 @@ class ClientResponseImpl extends Inspectable.Class implements ClientResponse.Htt
     return Effect.flatMap(this.text, (_) =>
       Effect.try({
         try: () => UrlParams.fromInput(new URLSearchParams(_)),
-        catch: (_) =>
+        catch: (cause) =>
           new Error.ResponseError({
             request: this.request,
             response: this,
             reason: "Decode",
-            error: _
+            cause
           })
       }))
   }
@@ -129,12 +129,12 @@ class ClientResponseImpl extends Inspectable.Class implements ClientResponse.Htt
   get formData(): Effect.Effect<FormData, Error.ResponseError> {
     return this.formDataBody ??= Effect.tryPromise({
       try: () => this.source.formData(),
-      catch: (_) =>
+      catch: (cause) =>
         new Error.ResponseError({
           request: this.request,
           response: this,
           reason: "Decode",
-          error: _
+          cause
         })
     }).pipe(Effect.cached, Effect.runSync)
   }
@@ -143,12 +143,12 @@ class ClientResponseImpl extends Inspectable.Class implements ClientResponse.Htt
   get arrayBuffer(): Effect.Effect<ArrayBuffer, Error.ResponseError> {
     return this.arrayBufferBody ??= Effect.tryPromise({
       try: () => this.source.arrayBuffer(),
-      catch: (_) =>
+      catch: (cause) =>
         new Error.ResponseError({
           request: this.request,
           response: this,
           reason: "Decode",
-          error: _
+          cause
         })
     }).pipe(Effect.cached, Effect.runSync)
   }